+++ src/Controller/InstagramController.php
@@ -2,23 +2,13 @@
 /**
  * This file is part of me-cms-instagram.
  *
- * me-cms-instagram is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
+ * Licensed under The MIT License
+ * For full copyright and license information, please see the LICENSE.txt
+ * Redistributions of files must retain the above copyright notice.
  *
- * me-cms-instagram is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with me-cms-instagram.  If not, see <http://www.gnu.org/licenses/>.
- *
- * @author      Mirko Pagliai <mirko.pagliai@gmail.com>
- * @copyright   Copyright (c) 2016, Mirko Pagliai for Nova Atlantis Ltd
- * @license     http://www.gnu.org/licenses/agpl.txt AGPL License
- * @link        http://git.novatlantis.it Nova Atlantis Ltd
+ * @copyright   Copyright (c) Mirko Pagliai
+ * @link        https://github.com/mirko-pagliai/me-cms-instagram
+ * @license     https://opensource.org/licenses/mit-license.php MIT License
  */
 namespace MeCmsInstagram\Controller;
 
@@ -80,7 +70,7 @@ class InstagramController extends AppController
     public function index($id = null)
     {
         //Sets initial cache name
-        $cache = sprintf('index_limit_%s', getConfig('default.photos'));
+        $cache = sprintf('index_limit_%s', getConfigOrFail('default.photos'));
 
         //Adds the request ID ("Next ID" for Istangram) to the cache name
         if (!empty($id)) {
@@ -92,7 +82,7 @@ class InstagramController extends AppController
 
         //If the data are not available from the cache
         if (empty($photos)) {
-            list($photos, $nextId) = $this->Instagram->recent($id, getConfig('default.photos'));
+            list($photos, $nextId) = $this->Instagram->recent($id, getConfigOrFail('default.photos'));
 
             Cache::write($cache, [$photos, $nextId], 'instagram');
         }