+++ src/View/Cell/PhotosWidgetsCell.php
@@ -2,23 +2,13 @@
 /**
  * This file is part of me-cms-instagram.
  *
- * me-cms-instagram is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
+ * Licensed under The MIT License
+ * For full copyright and license information, please see the LICENSE.txt
+ * Redistributions of files must retain the above copyright notice.
  *
- * me-cms-instagram is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with me-cms-instagram.  If not, see <http://www.gnu.org/licenses/>.
- *
- * @author      Mirko Pagliai <mirko.pagliai@gmail.com>
- * @copyright   Copyright (c) 2016, Mirko Pagliai for Nova Atlantis Ltd
- * @license     http://www.gnu.org/licenses/agpl.txt AGPL License
- * @link        http://git.novatlantis.it Nova Atlantis Ltd
+ * @copyright   Copyright (c) Mirko Pagliai
+ * @link        https://github.com/mirko-pagliai/me-cms-instagram
+ * @license     https://opensource.org/licenses/mit-license.php MIT License
  */
 namespace MeCmsInstagram\View\Cell;
 
@@ -65,7 +55,7 @@ class PhotosWidgetsCell extends Cell
      * @uses MeCmsInstagram\Utility\Instagram::recent()
      * @uses $Instagram
      */
-    protected function _latest($limit = 12)
+    protected function getLatest($limit = 12)
     {
         //Sets the cache name
         $cache = sprintf('widget_latest_%s', $limit);
@@ -87,7 +77,7 @@ class PhotosWidgetsCell extends Cell
      * Latest widget
      * @param int $limit Limit
      * @return void
-     * @uses _latest()
+     * @uses getLatest()
      */
     public function latest($limit = 1)
     {
@@ -96,7 +86,7 @@ class PhotosWidgetsCell extends Cell
             return;
         }
 
-        $photos = $this->_latest($limit);
+        $photos = $this->getLatest($limit);
 
         $this->set(compact('photos'));
     }
@@ -105,7 +95,7 @@ class PhotosWidgetsCell extends Cell
      * Random widget
      * @param int $limit Limit
      * @return void
-     * @uses _latest()
+     * @uses getLatest()
      */
     public function random($limit = 1)
     {
@@ -114,7 +104,7 @@ class PhotosWidgetsCell extends Cell
             return;
         }
 
-        $photos = collection($this->_latest())->sample($limit)->toArray();
+        $photos = collection($this->getLatest())->sample($limit)->toArray();
 
         $this->set(compact('photos'));
     }