+++ tests/TestCase/View/Cell/PhotosWidgetsCellTest.php
@@ -2,30 +2,20 @@
 /**
  * This file is part of me-cms-instagram.
  *
- * me-cms-instagram is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
+ * Licensed under The MIT License
+ * For full copyright and license information, please see the LICENSE.txt
+ * Redistributions of files must retain the above copyright notice.
  *
- * me-cms-instagram is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with me-cms-instagram.  If not, see <http://www.gnu.org/licenses/>.
- *
- * @author      Mirko Pagliai <mirko.pagliai@gmail.com>
- * @copyright   Copyright (c) 2016, Mirko Pagliai for Nova Atlantis Ltd
- * @license     http://www.gnu.org/licenses/agpl.txt AGPL License
- * @link        http://git.novatlantis.it Nova Atlantis Ltd
+ * @copyright   Copyright (c) Mirko Pagliai
+ * @link        https://github.com/mirko-pagliai/me-cms-instagram
+ * @license     https://opensource.org/licenses/mit-license.php MIT License
  */
 namespace MeCmsInstagram\Test\TestCase\View\Cell;
 
 use Cake\Cache\Cache;
-use Cake\TestSuite\TestCase;
 use MeCms\View\Helper\WidgetHelper;
 use MeCms\View\View\AppView as View;
+use MeTools\TestSuite\TestCase;
 
 /**
  * PhotosWidgetsCellTest class
@@ -58,32 +48,20 @@ class PhotosWidgetsCellTest extends TestCase
             $widgetClass = call_user_func_array([new WidgetHelper(new View), 'widget'], func_get_args());
 
             $widgetClass->Instagram = $this->getMockBuilder(get_class($widgetClass->Instagram))
-                ->setMethods(['_getRecentResponse'])
+                ->setMethods(['getRecentResponse'])
                 ->getMock();
 
+            $returnValue = file_get_contents(TEST_APP . 'examples' . DS . 'recent.json');
+
             if (in_array($this->getName(), ['testLatestNoPhotos', 'testRandomNoPhotos'])) {
-                $widgetClass->Instagram
-                    ->method('_getRecentResponse')
-                    ->will($this->returnValue(json_encode(['data' => []])));
-            } else {
-                $widgetClass->Instagram
-                    ->method('_getRecentResponse')
-                    ->will($this->returnValue(file_get_contents(TEST_APP . 'examples' . DS . 'recent.json')));
+                $returnValue = json_encode(['data' => []]);
             }
+
+            $widgetClass->Instagram->method('getRecentResponse')
+                ->will($this->returnValue($returnValue));
 
             return $widgetClass;
         }));
-    }
-
-    /**
-     * Teardown any static object changes and restore them
-     * @return void
-     */
-    public function tearDown()
-    {
-        parent::tearDown();
-
-        unset($this->Widget);
     }
 
     /**